backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
graphene-django>=3.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pandas>=2.0
Faker>=20.0
django-graphql-jwt==0.3.4